Split up event bus module.
