+++ engine/lib/phx/src/engine/mod.rs
@@ -1,7 +1,5 @@
 mod engine;
-mod event_bus;
 mod main_loop;
 
 pub use engine::*;
-pub use event_bus::*;
 pub use main_loop::*;
