+++ engine/lib/phx/src/lib.rs
@@ -8,6 +8,7 @@
 pub mod audio;
 pub mod common;
 pub mod engine;
+pub mod event_bus;
 pub mod error;
 pub mod input;
 pub mod logging;
